Move AbsenceTimeConfiguration into calendar package
closes #4976
