+++ src/main/java/org/synyx/urlaubsverwaltung/calendarintegration/CalendarSyncService.java
@@ -4,7 +4,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import org.synyx.urlaubsverwaltung.absence.AbsenceTimeConfiguration;
 import org.synyx.urlaubsverwaltung.application.application.Application;
 import org.synyx.urlaubsverwaltung.application.application.ApplicationAllowedEvent;
 import org.synyx.urlaubsverwaltung.application.application.ApplicationAllowedTemporarilyEvent;
@@ -14,6 +13,7 @@ import org.synyx.urlaubsverwaltung.application.application.ApplicationDeletedEve
 import org.synyx.urlaubsverwaltung.application.application.ApplicationRejectedEvent;
 import org.synyx.urlaubsverwaltung.application.application.ApplicationRevokedEvent;
 import org.synyx.urlaubsverwaltung.application.application.ApplicationUpdatedEvent;
+import org.synyx.urlaubsverwaltung.calendar.AbsenceTimeConfiguration;
 import org.synyx.urlaubsverwaltung.calendar.CalendarAbsence;
 import org.synyx.urlaubsverwaltung.settings.SettingsService;
 import org.synyx.urlaubsverwaltung.sicknote.sicknote.SickNote;
